feat: add PopupSession component
