+++ packages/core/src/components/templates/index.js
@@ -18,3 +18,4 @@
 
 export {default as AttachedFilesView} from './AttachedFilesView/AttachedFilesView';
 export {default as MailMessageView} from './MailMessageView/MailMessageView';
+export {default as PopupSession} from './PopupSession/PopupSession';
